+++ packages/nextra-theme-docs/src/components/back-to-top.tsx
@@ -39,7 +39,7 @@ export const BackToTop: FC<{
     >
       {children}
       <ArrowRightIcon
-        height="16"
+        height="1.1em"
         className="x:-rotate-90 x:border x:rounded-full x:border-current"
       />
     </Button>